WMS-98: Resstructuracion, creacion del layout y cambio a la carpeta paginas
