+++ src/main.js
@@ -1,5 +1,3 @@
-
-
 import { createApp } from 'vue'
 import App from './App.vue'
 import './assets/css/styles.css'
@@ -8,5 +6,3 @@ import router from './router';
 createApp(App)
   .use(router) 
   .mount('#app');
-
-
