+++ src/pages/Producto.vue
